fix(embeds): keep the steps boundary and split the card module
Two things CI caught that the local runs did not.

`provider_cards.rs` crossed the repo's 350-line limit for a changed file.
The card model and its two renderers move to `card_render.rs`, leaving the
provider module as the rules for recognising URLs; its siblings take the
model from there rather than reaching through the provider module for it.

The steps redesign dropped the panel #939 added to separate a steps preview
from the fence showing its own source, and that fix's regression test said
so. The list keeps its rail and loses the rounded filled box, but the
boundary comes back as rules above and below with a whisper of fill -- the
top-and-bottom-rule language the link previews already use -- so #939's
test passes unchanged.

The provider-card counts and the catalog pre-scan marker follow the article
platforms over to `.ox-ogp-card`, the link-preview stylesheet gets the same
did-it-ship assertion the provider frame has, and the kiosk matrix capture
is regenerated for the new chrome.
