+++ crates/ox_content_transform/src/media_embeds.rs
@@ -1,5 +1,6 @@
 mod apple_music;
 mod card_icons;
+mod card_render;
 #[cfg(test)]
 mod catalog_tests;
 mod document_cards;
